pyproject.toml
[tool.poetry]
name = "server_daangnmarket"
version = "0.1.0"
description = ""
authors = ["jam <nevvjann@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
djangorestframework = "^3.11.0"
django-filter = "^2.2.0"
markdown = "^3.2.1"
django-import-export = "^2.0.2"
psycopg2-binary = "^2.8.4"
Pillow = "^7.0.0"
requests = "^2.23.0"
supervisor = "^4.1.0"
gunicorn = "^20.0.4"
drf-yasg = "^1.17.1"
sentry-sdk = "^0.14.3"
ssv = "^0.1.1"
flex = "^6.14.1"
firebase-admin = "^4.0.0"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
